Get total loading tiles

+++ Cesium3DTilesSelection/include/Cesium3DTilesSelection/Tileset.h
@@ -153,7 +153,6 @@ public:
     return this->_overlays;
   }
 
-
   /**
    * @brief Updates this view but waits for all tiles that meet sse to finish
    * loading and ready to be rendered before returning the function. This method
@@ -194,7 +193,7 @@ public:
    */
   void notifyTileUnloading(Tile* pTile) noexcept;
 
-  const uint32_t getTilesetLoadingStatus() noexcept;
+  const size_t getTilesetLoadingStatus() noexcept;
 
   /**
    * @brief Loads a tile tree from a tileset.json file.
